Fix: cleanup frontend service
Fixes #40
