+++ CHANGES.md
@@ -5,12 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## Fixed
+### Fixed
 - Instagram Crawler (#24)
 - Telegram Channel username/title (#36)
 - Instagram jalali date (#39)
 
-[1.0.0] - 2021-07-13
+## [1.0.0] - 2021-07-13
 ### Added
 - Changelog file
 - Docker compose
